Updated Delete to be inline with other controllers

+++ Backend/WebAPI/Controllers/VitalsController.cs
@@ -113,6 +113,8 @@ namespace WebAPI.Controllers
         {
             try
             {
+                if (p_vitals == null)
+                    throw new Exception("Delete failed!");
                 _repo.Delete(p_vitals);
                 _repo.Save();
                 return Ok();
